Handle multiple jobs per yaml file
The job yaml files top level structure is a list, and there was no
reason it can't contain multiple jobs. But things were hard coded to
treat them as a single object in the list. This fixes that and makes
the job manager queue up a job for each entry in the yaml.