Add options to specify alternative run and deploy playbooks
This commit adds 2 new options (and supporting documentation) to enable
specifying alternative playbooks to use for deploying worker nodes and
running handbrake. Since every environment for running handbrakecloud is
likely to be different this will be useful to anyone to actually use it
outside of the private cloud in my closet.