Fix failure condition handling
This commit updates some of the error handling based on condtions I
encountered in my local testing. This includes adding a new optional
config parameter to retry failures if the run_handbrake phase fails. I
had a host go down on me during a run and this would have been useful
for that.